:recycle: Refactor into render submodules
